backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Google Sheets Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0

# AI & LangChain
langchain
langchain-openai
openai

# Data Processing
pandas==2.1.3
numpy==1.26.2

# API Support
pydantic==2.5.0
python-multipart==0.0.6

# CORS & Security
fastapi-cors==0.0.6
